feat: 2 pow 32 bytes guest memory address (#3104)
Adds support for 2^32 bytes memory address. Supersedes #2850.

Benchmark:
https://github.com/axiom-crypto/openvm-eth/actions/runs/31222063368

New design:
https://github.com/axiom-crypto/openvm-eth/actions/runs/32282531103/job/96164388705

Bench after MEM_BITS change:
https://github.com/axiom-crypto/openvm-eth/actions/runs/32427802054/job/96613241136

After fixing some codex audit findings:
https://github.com/axiom-crypto/openvm-eth/actions/runs/32431879330

Closes INT-8080

---------

+++ crates/sdk/src/lib.rs
@@ -402,7 +402,7 @@ where
         let elf_path =
             find_unique_executable(pkg_dir, target_dir, target_filter).map_err(SdkError::Other)?;
         let data = read(&elf_path)?;
-        Elf::decode(&data, MEM_SIZE as u32).map_err(SdkError::Other)
+        Elf::decode(&data, MEM_SIZE as u64).map_err(SdkError::Other)
     }
 
     /// Transpiler for transpiling RISC-V ELF to OpenVM executable.
@@ -442,7 +442,7 @@ where
             }),
             ExecutableInput::ElfFile(path) => {
                 let bytes = read(&path)?;
-                let elf = Elf::decode(&bytes, MEM_SIZE as u32)?;
+                let elf = Elf::decode(&bytes, MEM_SIZE as u64)?;
                 Ok(CompileInput {
                     executable: ExecutableFormat::Elf(elf),
                     #[cfg(feature = "rvr")]
